// src/grpc/proto/grpc/health/v1/health.ts
export interface HealthCheckRequest {
  service: string
}

export interface HealthCheckResponse {
  status: HealthCheckResponse_ServingStatus
}

export enum HealthCheckResponse_ServingStatus {
  UNKNOWN = 0,
  SERVING = 1,
  NOT_SERVING = 2,
  /** SERVICE_UNKNOWN - Used only by the Watch method. */
  SERVICE_UNKNOWN = 3,
  UNRECOGNIZED = -1,
}
// ...
export const HealthCheckRequest = {
  encode(
    message: HealthCheckRequest,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.service !== '') {
      writer.uint32(10).string(message.service)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): HealthCheckRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseHealthCheckRequest()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.service = reader.string()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): HealthCheckRequest {
    return { service: isSet(object.service) ? globalThis.String(object.service) : '' }
  },

  toJSON(message: HealthCheckRequest): unknown {
    const obj: any = {}
    if (message.service !== '') {
      obj.service = message.service
    }
    return obj
  },

  create<I extends Exact<DeepPartial<HealthCheckRequest>, I>>(
    base?: I,
  ): HealthCheckRequest {
    return HealthCheckRequest.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<HealthCheckRequest>, I>>(
    object: I,
  ): HealthCheckRequest {
    const message = createBaseHealthCheckRequest()
    message.service = object.service ?? ''
    return message
  },
}

function createBaseHealthCheckResponse(): HealthCheckResponse {
  return { status: 0 }
}

export const HealthCheckResponse = {
  encode(
    message: HealthCheckResponse,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.status !== 0) {
      writer.uint32(8).int32(message.status)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): HealthCheckResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseHealthCheckResponse()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.status = reader.int32() as any
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): HealthCheckResponse {
    return {
      status: isSet(object.status)
        ? healthCheckResponse_ServingStatusFromJSON(object.status)
        : 0,
    }
  },

  toJSON(message: HealthCheckResponse): unknown {
    const obj: any = {}
    if (message.status !== 0) {
      obj.status = healthCheckResponse_ServingStatusToJSON(message.status)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<HealthCheckResponse>, I>>(
    base?: I,
  ): HealthCheckResponse {
    return HealthCheckResponse.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<HealthCheckResponse>, I>>(
    object: I,
  ): HealthCheckResponse {
    const message = createBaseHealthCheckResponse()
    message.status = object.status ?? 0
    return message
  },
}
// ...
export interface HealthServer extends UntypedServiceImplementation {
  check: handleUnaryCall<HealthCheckRequest, HealthCheckResponse>
  watch: handleServerStreamingCall<HealthCheckRequest, HealthCheckResponse>
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }